PN-13608: Fixed check on invalid and not PEC addresses. Renamed default packages. Removed template classes. Fixed tests.
