--- config/application.properties
@@ -1,7 +1,0 @@
-logging.config=config/logback-local.xml
-
-# Runtime mode
-pn.env.runtime=DEVELOPMENT
-#pn.env.runtime=PROD
-
-
