+++ src/main/java/com/dummy/pec/library/conf/DummyPecSharedAutoConfiguration.java
@@ -1,0 +1,10 @@
+package com.dummy.pec.library.conf;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.util.ResourceUtils;
+
+@Configuration
+@PropertySource(ResourceUtils.CLASSPATH_URL_PREFIX + "application-shared-dummy.properties")
+public class DummyPecSharedAutoConfiguration {
+}
