+++ src/main/java/com/dummy/pec/library/type/PecType.java
@@ -1,0 +1,8 @@
+package com.dummy.pec.library.type;
+
+public enum PecType {
+    CONSEGNA,
+    ACCETTAZIONE,
+
+    NON_PEC;
+}
