--- src/main/java/it/pagopa/pn/template/TemplateApplication.java
@@ -1,27 +1,0 @@
-package it.pagopa.pn.template;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-public class TemplateApplication {
-
-
-    public static void main(String[] args) {
-        SpringApplication.run(TemplateApplication.class, args);
-    }
-
-
-    @RestController
-    @RequestMapping("/")
-    public static class RootController {
-
-        @GetMapping("/")
-        public String home() {
-            return "";
-        }
-    }
-}
