--- src/main/java/it/pagopa/pn/template/type/PecType.java
@@ -1,6 +1,0 @@
-package it.pagopa.pn.template.type;
-
-public enum PecType {
-    CONSEGNA,
-    ACCETTAZIONE;
-}
