+++ src/main/resources/application-shared-dummy.properties
@@ -1,0 +1,5 @@
+dummy.pec.min-delay-ms=100
+dummy.pec.max-delay-ms=500
+
+#elenco indirizzi in blacklist (separati da virgola)
+blacklisted.addresses=test1@test.it,test2@xxx.com,
