--- src/main/resources/application.properties
@@ -1,10 +1,0 @@
-logging.config=classpath:logback-base.xml
-
-pn.env.runtime=PROD
-spring.application.name=PN-TEMPLATE-MS-BE
-
-#elenco indirizzi in blacklist (separati da virgola)
-blacklisted.addresses=test1@test.it,test2@xxx.com,
-
-#Regular Expression (RFC 5322)
-regex.pattern=^[a-zA-Z0-9_!#$%&?*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$
