+++ src/test/java/com/dummy/pec/library/DummyPecTestApplication.java
@@ -1,0 +1,7 @@
+package com.dummy.pec.library;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DummyPecTestApplication {
+}
