--- src/test/java/it/pagopa/pn/template/rest/TemplateSampleApiControllerTest.java
@@ -1,19 +1,0 @@
-package it.pagopa.pn.template.rest;
-
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.fail;
-
-@SpringBootTest
-public class TemplateSampleApiControllerTest {
-
-
-    @Test
-    void getHttpHeadersMap() {
-        Assertions.assertTrue(true, "Not yet Implemented");
-
-    }
-}
