--- src/test/java/it/pagopa/pn/template/service/DummyPecTestApplication.java
@@ -1,7 +1,0 @@
-package it.pagopa.pn.template.service;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DummyPecTestApplication {
-}
