+++ src/test/resources/application-test.properties
@@ -1,0 +1,1 @@
+logging.config=classpath:logback-test.xml
